+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/components/maps/AirMapOverlay.java
@@ -22,6 +22,7 @@ public class AirMapOverlay extends AirMapFeature implements ImageReadable {
   private LatLngBounds bounds;
   private BitmapDescriptor iconBitmapDescriptor;
   private Bitmap iconBitmap;
+  private boolean tappable;
   private float zIndex;
   private float transparency;
 
@@ -60,6 +61,13 @@ public class AirMapOverlay extends AirMapFeature implements ImageReadable {
     this.mImageReader.setImage(uri);
   }
 
+  public void setTappable(boolean tapabble) {
+    this.tappable = tapabble;
+    if (groundOverlay != null) {
+      groundOverlay.setClickable(tappable);
+    }
+  }
+
 
   public GroundOverlayOptions getGroundOverlayOptions() {
     if (this.groundOverlayOptions == null) {
@@ -72,14 +80,19 @@ public class AirMapOverlay extends AirMapFeature implements ImageReadable {
     if (this.groundOverlayOptions != null) {
       return this.groundOverlayOptions;
     }
+    GroundOverlayOptions options = new GroundOverlayOptions();
     if (this.iconBitmapDescriptor != null) {
-      GroundOverlayOptions options = new GroundOverlayOptions();
       options.image(iconBitmapDescriptor);
-      options.positionFromBounds(bounds);
-      options.zIndex(zIndex);
-      return options;
+    } else {
+      // add stub image to be able to instantiate the overlay
+      // and store a reference to it in MapView
+      options.image(BitmapDescriptorFactory.defaultMarker());
+      // hide overlay until real image gets added
+      options.visible(false);
     }
-    return null;
+    options.positionFromBounds(bounds);
+    options.zIndex(zIndex);
+    return options;
   }
 
   @Override
@@ -92,7 +105,7 @@ public class AirMapOverlay extends AirMapFeature implements ImageReadable {
     GroundOverlayOptions groundOverlayOptions = getGroundOverlayOptions();
     if (groundOverlayOptions != null) {
       this.groundOverlay = map.addGroundOverlay(groundOverlayOptions);
-      this.groundOverlay.setClickable(true);
+      this.groundOverlay.setClickable(this.tappable);
     } else {
       this.map = map;
     }
@@ -123,8 +136,9 @@ public class AirMapOverlay extends AirMapFeature implements ImageReadable {
   public void update() {
     this.groundOverlay = getGroundOverlay();
     if (this.groundOverlay != null) {
+      this.groundOverlay.setVisible(true);
       this.groundOverlay.setImage(this.iconBitmapDescriptor);
-      this.groundOverlay.setClickable(true);
+      this.groundOverlay.setClickable(this.tappable);
     }
   }