+++ tools/expotools/src/commands/UpdateVendoredModule.ts
@@ -52,16 +52,17 @@ const vendoredModulesConfig: { [key: string]: VendoredModuleConfig } = {
     semverPrefix: '~',
     steps: [
       {
-        sourceIosPath: 'ios',
-        targetIosPath: 'Api/Components/GestureHandler',
-        sourceAndroidPath: 'android/src/main/java/com/swmansion/gesturehandler/react',
-        targetAndroidPath: 'modules/api/components/gesturehandler/react',
+        sourceAndroidPath: 'android/lib/src/main/java/com/swmansion/gesturehandler',
+        targetAndroidPath: 'modules/api/components/gesturehandler',
         sourceAndroidPackage: 'com.swmansion.gesturehandler',
         targetAndroidPackage: 'versioned.host.exp.exponent.modules.api.components.gesturehandler',
       },
       {
-        sourceAndroidPath: 'android/lib/src/main/java/com/swmansion/gesturehandler',
-        targetAndroidPath: 'modules/api/components/gesturehandler',
+        recursive: true,
+        sourceIosPath: 'ios',
+        targetIosPath: 'Api/Components/GestureHandler',
+        sourceAndroidPath: 'android/src/main/java/com/swmansion/gesturehandler/react',
+        targetAndroidPath: 'modules/api/components/gesturehandler/react',
         sourceAndroidPackage: 'com.swmansion.gesturehandler',
         targetAndroidPackage: 'versioned.host.exp.exponent.modules.api.components.gesturehandler',
       },
@@ -133,7 +134,7 @@ const vendoredModulesConfig: { [key: string]: VendoredModuleConfig } = {
     steps: [
       {
         sourceIosPath: 'ios',
-        targetIosPath: 'Api',
+        targetIosPath: 'Api/ViewShot',
         sourceAndroidPath: 'android/src/main/java/fr/greweb/reactnativeviewshot',
         targetAndroidPath: 'modules/api/viewshot',
         sourceAndroidPackage: 'fr.greweb.reactnativeviewshot',
@@ -196,6 +197,7 @@ const vendoredModulesConfig: { [key: string]: VendoredModuleConfig } = {
         targetIosPath: 'Api/Components/GoogleMaps',
       },
       {
+        recursive: true,
         sourceIosPath: 'lib/ios/AirMaps',
         targetIosPath: 'Api/Components/Maps',
         sourceAndroidPath: 'lib/android/src/main/java/com/airbnb/android/react/maps',
@@ -221,7 +223,7 @@ const vendoredModulesConfig: { [key: string]: VendoredModuleConfig } = {
     ],
   },
   'react-native-webview': {
-    repoUrl: 'https://github.com/expo/react-native-webview.git',
+    repoUrl: 'https://github.com/react-native-community/react-native-webview.git',
     installableInManagedApps: true,
     steps: [
       {